pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "connes-cvs"
version = "0.1.0"
description = "First public implementation of the Connes-van Suijlekom Galerkin matrix for the Riemann Hypothesis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Akiva Groskin" },
]
keywords = [
    "riemann-hypothesis",
    "spectral-theory",
    "number-theory",
    "connes",
    "weil-positivity",
    "galerkin",
    "arbitrary-precision",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "mpmath>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
fast = [
    "python-flint>=0.5.0",
]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
    "gmpy2>=2.1.0",
]
all = [
    "python-flint>=0.5.0",
    "gmpy2>=2.1.0",
]

[project.urls]
Homepage = "https://github.com/akivagroskin/connes-cvs"
Documentation = "https://github.com/akivagroskin/connes-cvs#readme"
Repository = "https://github.com/akivagroskin/connes-cvs"
Issues = "https://github.com/akivagroskin/connes-cvs/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that take >60 seconds (e.g., full matrix construction)",
]
